+++ implementations/delivery/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @cotal-ai/delivery
 
+## 0.14.8
+
+### Patch Changes
+
+- Updated dependencies [84f6200]
+  - @cotal-ai/core@0.14.8
+  - @cotal-ai/workspace@0.14.8
+
+## 0.14.7
+
+### Patch Changes
+
+- Updated dependencies [12ad5e3]
+  - @cotal-ai/workspace@0.14.7
+  - @cotal-ai/core@0.14.7
+
 ## 0.14.6
 
 ### Patch Changes